% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afr.gwas.R
\docType{data}
\name{afr.gwas}
\alias{afr.gwas}
\title{GWAS data sets}
\format{
An object of class \code{list} of length 7.
}
\usage{
afr.gwas
}
\description{
A list of 7 population study files of African ancestry.
Each GWA study file has mandatory column headers: MARKERNAME, EA, NEA, EAF, CHROMOSOME,
POSITION, BETA AND SE.
}
\keyword{datasets}
