% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{check_eaf}
\alias{check_eaf}
\title{Check EA and NEA.}
\usage{
check_eaf(ssp)
}
\arguments{
\item{ssp}{A list of all GWAS}
}
\value{
A list of all GWAS which would be aligned to the same reference allele.
}
\description{
Check whether EA and NEA of each genetic variant follow the same reference allele.
}
