% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filt_sum_stat.R
\name{check_uniq}
\alias{check_uniq}
\title{Check unique}
\usage{
check_uniq(ssp, ncores)
}
\arguments{
\item{ssp}{A dataframe of one GWAS}

\item{ncores}{The number of cores which would be used for running in parallel.}
}
\value{
A list of all study files in which genetic variants shown more than once would be filtered out.
}
\description{
Function: Check whether there are some genetic variants shown more than once in each cohort.
}
\author{
Siru Wang
}
