% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{dist_pop}
\alias{dist_pop}
\title{Euclidean distance matrix}
\usage{
dist_pop(X)
}
\arguments{
\item{X}{MAF of selected variants with MAF>5\% across all cohorts.}
}
\value{
Output the Euclidean distance matrix.
}
\description{
Calculate the Euclidean distance matrix.
}
