% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{double_center}
\alias{double_center}
\title{Double-center distance matrix}
\usage{
double_center(D)
}
\arguments{
\item{D}{The Euclidean distance matrix.}
}
\value{
The multi-dimensional scaling of the Euclidean distance matrix.
}
\description{
Calculate multi-dimensional scaling of the distance matrix.
}
