% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{getN_AverageEAF}
\alias{getN_AverageEAF}
\title{Calculate the weighted MAF.}
\usage{
getN_AverageEAF(ssp, ssf, ncores)
}
\arguments{
\item{ssp}{A list of all GWAS.}

\item{ssf}{A vector including the names of the filtered genetic variants across all cohorts .}

\item{ncores}{The the number of cores which would be used for running in parallel.}
}
\value{
The output is the weighted MAF
}
\description{
Calculate the weighted MAF.
}
\author{
Siru Wang
}
