% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_input.R
\docType{data}
\name{gwas_input}
\alias{gwas_input}
\title{Categorized GWAS data}
\format{
An object of class \code{list} of length 5.
}
\usage{
gwas_input
}
\description{
A list of length 7. The details of 7 objects are shown in the following:
beta_pop object: BETA values across 7 cohorts.
invse2_pop object: inverse of variance across 7 cohorts.
marker_inf_pop object: basic information of each genetic variant, which consisting of MARKERNAME, CHROMOSOME, EA and NEA.
corhort_count_filt object: the frequecies of each genetic variant present across 7 cohorts.
sel_gene object: the genetic variants which are filtered out.
pcs object: the axes of genetic variation which would be used in the meta-regression model.
env object: the smoking proportions across 7 cohorts
}
\keyword{datasets}
