% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_linear_model.R
\name{MR_mega_lrw}
\alias{MR_mega_lrw}
\title{Linear regression model with weights for MR-MEGA approach}
\usage{
MR_mega_lrw(dta, qt = TRUE, ncores, pcCount)
}
\arguments{
\item{dta}{A list containing BETA, SE^{-2},PCs, cohort_count_filt which may be used for Bayesian factor.}

\item{qt}{Whether the trait is quantitative or not. This method is only used for quantitative traits.}

\item{ncores}{The the number of cores which would be used for running in parallel.}

\item{pcCount}{The number of axes of genetic variation.}
}
\value{
Output a file containing names of genetic variants, estimated coefficients, standard errors, chisq value of the association, the number of degrees of freedom of the association, p-value of the association, chisq value of the heterogeneity due to different ancestry, ndf of the heterogeneity due to different ancestry, p-value of the heterogeneity due to different ancestry, chisq value of the residual heterogeneity, ndf of the residual heterogeneity,  p-value of the residual heterogeneity.
}
\description{
Fit the meta-regression model using the normalized statistic data across all cohorts.
}
\details{
The meta-regression model was proposed by \insertCite{magi2017trans;textual}{env.MRmega}.
}
\references{
\insertRef{magi2017trans}{env.MRmega}
}
\author{
Siru Wang
}
