% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_PC_generate.R
\name{MR_mega_MDS}
\alias{MR_mega_MDS}
\title{Calculate MDS of the Euclidean distance matrix}
\usage{
MR_mega_MDS(data_name, batch_loc, qt = TRUE, pcCount = 2, ncores)
}
\arguments{
\item{data_name}{A file containing all GWAS file names and the corresponding locations.}

\item{batch_loc}{The location of the batches.}

\item{qt}{Whether the trait is quantitative or not. This method is only used for quantitative traits.}

\item{pcCount}{The number of axes of genetic variation. Default the number of PCs is 1 and the limitation of pcCount should be pcCount+envCount<n_cohort-2}

\item{ncores}{The number of cores which would be used for running in parallel.}
}
\value{
null
}
\description{
Input the reference panel and calculate the associated MDS of the Euclidean distance matrix.
}
\details{
The meta-regression model was proposed by \insertCite{magi2017trans;textual}{env.MRmega}.
}
\references{
\insertRef{magi2017trans}{env.MRmega}
}
\author{
Siru Wang
}
