% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filt_sum_stat.R
\name{MR_mega_normalize}
\alias{MR_mega_normalize}
\title{MR_mega_normalize}
\usage{
MR_mega_normalize(data_name, qt = TRUE, ncores)
}
\arguments{
\item{data_name}{A vector of data name indicating the specific location of the summary statistic data of one GWAS.}

\item{qt}{Whether the trait is quantitative or not. This method is only used for quantitative traits.}

\item{ncores}{The number of cores which would be used for running in parallel.}
}
\value{
Output the normalized summary statistic data across different cohorts
}
\description{
Normalize the inputted file of summary statistic data across different cohorts, which includes
deleting all genetic variants with NA's and filtering out genetic variants shown more once.
}
\author{
Siru Wang
}
