% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{MR_mega_prep}
\alias{MR_mega_prep}
\title{Prepare data for MR-mega approach and environment-adjusted MR-MEGA approach (mr-mega)}
\usage{
MR_mega_prep(
  sum_stat_pop,
  n_cohort = NULL,
  usefor = "mr-mega",
  qt = TRUE,
  pcCount = 2,
  calpc = FALSE,
  envCount = NULL,
  ncores
)
}
\arguments{
\item{sum_stat_pop}{A list of all GWAS file data.}

\item{n_cohort}{The number of the cohorts.}

\item{usefor}{"mr-mega" or "env-mr-mega". The default option is "mr-mega".}

\item{qt}{Whether the trait is quantitative or not. This method is only used for quantitative traits.}

\item{pcCount}{The number of axes of genetic variation. Default the number of PCs is 1. The limitation of \code{pcCount} should be \code{pcCount}+\code{envCount}<\code{n_cohort}-2 for env-adjusted MR MEGA approach and
The limitation of \code{pcCount} should be \code{pcCount}<\code{n_cohort}-2 for MR MEGA approach.}

\item{calpc}{Logical. whether need to calculate PCs based on the inputted GWAS data.}

\item{envCount}{The number of environment covariates. Default the number of environment covariates is null.The limitation of envCount should be \code{pcCount}+\code{envCount}<\code{n_cohort}-2 for env-adjusted MR MEGA approach.}

\item{ncores}{The the number of cores which would be used for running in parallel.}
}
\value{
A list containing BETA, inverse of squared SE, PCs if required, marker information, cohort_count_filt and filtered-out genetic variants.
}
\description{
Prepare data for MR-mega approach and environment-adjusted MR-MEGA approach (mr-mega)
}
\seealso{
\code{\link{MR_mega_MDS}} is the function to calculate the MDS of Euclidean distance matrix and the associated axes of genetic variation.
}
\author{
Siru Wang
}
