% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{readInputfile}
\alias{readInputfile}
\title{read inputted file}
\usage{
readInputfile(data_name, qt = TRUE)
}
\arguments{
\item{data_name}{A file involving all names of summary statistic data across different cohorts}

\item{qt}{Whether the trait is quantitative or not. This method is only used for quantitative traits.}
}
\value{
A list of of all study files.
}
\description{
Read the inputted of file including all names of summary statistic data across different cohorts and select useful columns which
would be used for meta analysis.
}
