% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocedure_sum_stat.R
\name{sel_MafChrPos}
\alias{sel_MafChrPos}
\title{Select the quantified genetic variants with the specified threshold of MAF, chromosome and position.}
\usage{
sel_MafChrPos(ssp, sp)
}
\arguments{
\item{ssp}{A list of all GWAS.}

\item{sp}{A vector including the names of the genetic variants across all cohorts.}
}
\value{
A dataframe including the names of the genetic variants across all cohorts, the corresponding chromosomes, positions,ID and EAF across all cohorts.
}
\description{
Select the quantified genetic variants with the specified threshold of MAF, chromosome and position.
}
\author{
Siru Wang
}
