% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.env.R
\docType{data}
\name{smoke.env}
\alias{smoke.env}
\title{The smoking proportions across the 7 populations.}
\format{
An object of class \code{data.frame} with 4000 rows and 7 columns.
}
\usage{
smoke.env
}
\description{
A data frame of size 4000 times 7, each element refers to
the smoking status for each individual in the corresponding cohort.
}
\keyword{datasets}
